Metapopulation capacity

Metapopulation capacity (Hanski and Ovaskainen 2000) is a relative measure of the ability of a spatially explicit landscape to support a metapopulation of a species. This function implements the modification proposed by Schnell et al. (2013) that includes self colonization.

Usage

meta_capacity(x, a, f, ex = 0.5, self = TRUE, patch_mc = FALSE, ...)

Arguments

x

interpatch distance matrix or patch_config object.

a

numeric; patch areas in the same order as the distance matrix. Only used if `x“ is a distance matrix.

f

function; dispersal kernel function taking a distance, d, and returning the probability density of dispersing that distance.

ex

numeric; extinction-area parameter, where extinction rate is given by \(E / A^ex\) for patch size \(A\).

self

logical; whether patches can recolonize themselves, this correction was proposed by Schnell et al. (2013) because the original favours networks of small patches over large contiguous patches

patch_mc

logical; whether to return the patch-level contributions to the metapopulation capacity.

...

additional parameters passed to the dispersal survival function.

Value

The metapopulation capacity or, if patch_mc = TRUE, a list containing the metapopulation capacity (capacity) and the patch contributions (patch_mc).

References

Hanski, I, & O Ovaskainen (2000). The metapopulation capacity of a fragmented landscape. Nature, 404(6779), 755. doi:10.1038/35008063

Schnell, JK, GM Harris, SL Pimm, & GJ Russell (2013). Estimating Extinction Risk with Metapopulation Models of Large‐Scale Fragmentation. Conservation Biology, 27(3), 520-530. doi:10.1111/cobi.12047

Examples

d <- matrix(c(0, 2, 4, 2, 0, 3, 4, 3, 0), nrow = 3)
a <- c(1, 1.5, 0.5)
f <- dispersal_negexp(1)
meta_capacity(d, a, f)
#> [1] 1.876679

r <- raster::raster(nrows = 10, ncols = 10, crs = "+proj=laea")
r[] <- round(runif(raster::ncell(r)) * 0.7)
pc <- patch_config(r, "m")
f <- dispersal_negexp(1 / 100)
meta_capacity(pc, f = f)
#> [1] 295782.6