Metapopulation capacity of a candidate reserve

For use in conservation prioritization exercises, this function computes the metapopulation capacities for a suite of species within a selected set of planning units.

mc_reserve(pu, x, f, threshold = 0, ...)

Arguments

pu
raster::RasterStack, sp::SpatialPolygonsDataFrame, or sf::sf object; planning units and representation levels of features as layers (raster inputs) or columns (vector inputs).

x
logical or binary; indicator variables specifying which planning units are included.

f
named list of dispersal survival functions with names matching the layer/column names in pu.

threshold
numeric; representation level threshold required to consider a species present in a planning unit.

...
additional arguments passed on to meta_capacity().

Value

A numeric vector of metapopulation capacities for each species.

Examples

# generate data
r <- raster::raster(nrows = 10, ncols = 10, crs = "+proj=aea",
                    vals = sample(0:1, 100, replace = TRUE))
s <- raster::stack(r, r, r)
s[[2]][] <- sample(0:1, 100, replace = TRUE, prob = c(0.6, 0.4))
s[[3]][] <- sample(0:1, 100, replace = TRUE, prob = c(0.8, 0.2))
names(s) <- c("a", "b", "c")
selected <- sample(c(FALSE, TRUE), 100, replace = TRUE, prob = c(0.7, 0.3))
disp_f <- list(a = dispersal_negexp(1 / 0.01),
               b = dispersal_negexp(1 / 0.005),
               c = dispersal_negexp(1 / 0.02))
mc_reserve(s, selected, disp_f)

#>            a            b            c 
#> 0.0009592509 0.0000858041 0.0001429214 

# vector data
features <- raster::rasterToPolygons(s)
features <- sf::st_as_sf(features)
mc_reserve(features, selected, disp_f)

#>            a            b            c 
#> 2.447702e-04 6.322648e-05 1.026024e-04