Calculate interpatch distances and areas

Given a set of patches in raster or vector format, calculate the areas and interpatch distances.

patch_config(x, units = c("m", "km"))

Arguments

x
patch network in raster or vector format, see Details.

units
character; whether distances and areas should be in meters or kilometers.

Value

A patch_config object with two elements: the interpatch distances (distances) and patch areas (areas).

Details

If the patch network is provided as a Raster object, cells included in the network should be 1, and background cells should be 0 or NA. Patches can also be provided as polygons from either the sf or sp packages. Spatial data must be provided in projecred coordinates with units of meters.

Examples

# raster
r <- raster::raster(nrows = 10, ncols = 10, crs = "+proj=aea")
r[] <- round(runif(raster::ncell(r)) * 0.7)
patch_config(r, units = "m")

#> $areas
#>  [1] 2592  648 1296 1296  648  648  648 3240  648  648 3240 1944  648
#> 
#> $distances
#>            [,1]      [,2]     [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,]   0.00000 113.84200 108.0000 129.7998 145.1206 108.0000 108.00000
#>  [2,] 113.84200   0.00000  72.0000 180.0000   0.0000 250.0640 113.84200
#>  [3,] 108.00000  72.00000   0.0000  72.0000 108.0000 165.9518 180.00000
#>  [4,] 129.79985 180.00000  72.0000   0.0000 216.0000 126.0000 274.16783
#>  [5,] 145.12064   0.00000 108.0000 216.0000   0.0000 290.2413 126.00000
#>  [6,] 108.00000 250.06399 165.9518 126.0000 290.2413   0.0000 288.00000
#>  [7,] 108.00000 113.84200 180.0000 274.1678 126.0000 288.0000   0.00000
#>  [8,]  18.00000 129.79985  72.0000  54.0000 169.8117  18.0000 180.89776
#>  [9,]  18.00000  80.49845  72.0000 129.7998 115.2562 148.4318 109.48973
#> [10,]  80.49845  54.00000 120.7477 222.6477  72.0000 257.7208  36.00000
#> [11,]  64.89992  18.00000  18.0000 109.4897  36.0000 169.8117  64.89992
#> [12,] 101.82338 180.89776  74.2159  18.0000 218.9795  72.0000 262.08396
#> [13,]  90.00000 108.00000   0.0000  36.0000 145.1206 129.7998 201.24612
#>            [,8]      [,9]     [,10]     [,11]    [,12]     [,13]
#>  [1,]  18.00000  18.00000  80.49845  64.89992 101.8234  90.00000
#>  [2,] 129.79985  80.49845  54.00000  18.00000 180.8978 108.00000
#>  [3,]  72.00000  72.00000 120.74767  18.00000  74.2159   0.00000
#>  [4,]  54.00000 129.79985 222.64770 109.48973  18.0000  36.00000
#>  [5,] 169.81166 115.25624  72.00000  36.00000 218.9795 145.12064
#>  [6,]  18.00000 148.43180 257.72078 169.81166  72.0000 129.79985
#>  [7,] 180.89776 109.48973  36.00000  64.89992 262.0840 201.24612
#>  [8,]   0.00000  36.00000 144.00000  50.91169   0.0000  50.91169
#>  [9,]  36.00000   0.00000  72.00000  36.00000 113.8420  64.89992
#> [10,] 144.00000  72.00000   0.00000   0.00000 216.7487 148.43180
#> [11,]  50.91169  36.00000   0.00000   0.00000 108.0000  36.00000
#> [12,]   0.00000 113.84200 216.74870 108.00000   0.0000  36.00000
#> [13,]  50.91169  64.89992 148.43180  36.00000  36.0000   0.00000
#> 
#> attr(,"class")
#> [1] "patch_config"

# polygon
p_sp <- raster::rasterToPolygons(r, dissolve = TRUE)
p_sp <- p_sp[p_sp$layer == 1, ]
p_sp <- sp::disaggregate(p_sp)
patch_config(p_sp, units = "km")

#> $areas
#>  [1] 0.001296 0.000648 0.001296 0.000648 0.000648 0.000648 0.002592 0.000648
#>  [9] 0.000648 0.003240 0.001944 0.000648 0.003240
#> 
#> $distances
#>            [,1]      [,2]      [,3]       [,4]       [,5]       [,6]       [,7]
#>  [1,] 0.0000000 0.2160000 0.0720000 0.18000000 0.03600000 0.27416783 0.12979985
#>  [2,] 0.2160000 0.0000000 0.1080000 0.00000000 0.14512064 0.12600000 0.14512064
#>  [3,] 0.0720000 0.1080000 0.0000000 0.07200000 0.00000000 0.18000000 0.10800000
#>  [4,] 0.1800000 0.0000000 0.0720000 0.00000000 0.10800000 0.11384200 0.11384200
#>  [5,] 0.0360000 0.1451206 0.0000000 0.10800000 0.00000000 0.20124612 0.09000000
#>  [6,] 0.2741678 0.1260000 0.1800000 0.11384200 0.20124612 0.00000000 0.10800000
#>  [7,] 0.1297998 0.1451206 0.1080000 0.11384200 0.09000000 0.10800000 0.00000000
#>  [8,] 0.1260000 0.2902413 0.1659518 0.25006399 0.12979985 0.28800000 0.10800000
#>  [9,] 0.1297998 0.1152562 0.0720000 0.08049845 0.06489992 0.10948973 0.01800000
#> [10,] 0.1094897 0.0360000 0.0180000 0.01800000 0.03600000 0.06489992 0.06489992
#> [11,] 0.0180000 0.2189795 0.0742159 0.18089776 0.03600000 0.26208396 0.10182338
#> [12,] 0.2226477 0.0720000 0.1207477 0.05400000 0.14843180 0.03600000 0.08049845
#> [13,] 0.0540000 0.1698117 0.0720000 0.12979985 0.05091169 0.18089776 0.01800000
#>            [,8]       [,9]      [,10]     [,11]      [,12]      [,13]
#>  [1,] 0.1260000 0.12979985 0.10948973 0.0180000 0.22264770 0.05400000
#>  [2,] 0.2902413 0.11525624 0.03600000 0.2189795 0.07200000 0.16981166
#>  [3,] 0.1659518 0.07200000 0.01800000 0.0742159 0.12074767 0.07200000
#>  [4,] 0.2500640 0.08049845 0.01800000 0.1808978 0.05400000 0.12979985
#>  [5,] 0.1297998 0.06489992 0.03600000 0.0360000 0.14843180 0.05091169
#>  [6,] 0.2880000 0.10948973 0.06489992 0.2620840 0.03600000 0.18089776
#>  [7,] 0.1080000 0.01800000 0.06489992 0.1018234 0.08049845 0.01800000
#>  [8,] 0.0000000 0.14843180 0.16981166 0.0720000 0.25772078 0.01800000
#>  [9,] 0.1484318 0.00000000 0.03600000 0.1138420 0.07200000 0.03600000
#> [10,] 0.1698117 0.03600000 0.00000000 0.1080000 0.00000000 0.05091169
#> [11,] 0.0720000 0.11384200 0.10800000 0.0000000 0.21674870 0.00000000
#> [12,] 0.2577208 0.07200000 0.00000000 0.2167487 0.00000000 0.14400000
#> [13,] 0.0180000 0.03600000 0.05091169 0.0000000 0.14400000 0.00000000
#> 
#> attr(,"class")
#> [1] "patch_config"