Calculate interpatch distances and areas

Given a set of patches in raster or vector format, calculate the areas and interpatch distances.

patch_config(x, units = c("km", "m"))

Arguments

x
patch network in raster or vector format, see Details.

units
character; whether distances and areas should be in meters or kilometers.

Value

A patch_config object with two elements: the interpatch distances (distances) and patch areas (areas).

Details

If the patch network is provided as a Raster object, cells included in the network should be 1, and background cells should be 0 or NA. Patches can also be provided as polygons from either the sf or sp packages.

Examples

# raster
r <- raster::raster(nrows = 10, ncols = 10, crs = "+proj=aea")
r[] <- round(runif(raster::ncell(r)) * 0.7)
patch_config(r, units = "m")

#> $areas
#> [1]  648  648 1944 7776  648 3240 1944  648  648
#> 
#> $distances
#>           [,1]     [,2]      [,3]     [,4]      [,5]     [,6]     [,7]     [,8]
#>  [1,]   0.0000  36.0000 144.00000 36.00000 183.56470 101.8234 191.3426 144.0000
#>  [2,]  36.0000   0.0000  72.00000 54.00000 113.84200 160.9969 145.1206 144.0000
#>  [3,] 144.0000  72.0000   0.00000 64.89992  18.00000 241.4953 108.0000 165.9518
#>  [4,]  36.0000  54.0000  64.89992  0.00000  50.91169  18.0000  18.0000  18.0000
#>  [5,] 183.5647 113.8420  18.00000 50.91169   0.00000 262.0840  72.0000 169.8117
#>  [6,] 101.8234 160.9969 241.49534 18.00000 262.08396   0.0000 216.0000  72.0000
#>  [7,] 191.3426 145.1206 108.00000 18.00000  72.00000 216.0000   0.0000 108.0000
#>  [8,] 144.0000 144.0000 165.95180 18.00000 169.81166  72.0000 108.0000   0.0000
#>  [9,] 160.9969 144.0000 131.04198 18.00000 115.25624 144.0000  36.0000  36.0000
#>           [,9]
#>  [1,] 160.9969
#>  [2,] 144.0000
#>  [3,] 131.0420
#>  [4,]  18.0000
#>  [5,] 115.2562
#>  [6,] 144.0000
#>  [7,]  36.0000
#>  [8,]  36.0000
#>  [9,]   0.0000
#> 
#> attr(,"class")
#> [1] "patch_config"

# polygon
p_sp <- raster::rasterToPolygons(r, dissolve = TRUE)
p_sp <- p_sp[p_sp$layer == 1, ]
p_sp <- sp::disaggregate(p_sp)
patch_config(p_sp, units = "km")

#> $areas
#>  [1] 0.000648 0.000648 0.001296 0.001944 0.001296 0.000648 0.001944 0.000648
#>  [9] 0.000648 0.001944 0.001944 0.000648 0.000648 0.000648 0.000648 0.000648
#> [17] 0.000648 0.000648
#> 
#> $distances
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.00000000 0.03600000 0.07200000 0.10800000 0.01800000 0.16981166
#>  [2,] 0.03600000 0.00000000 0.14400000 0.03600000 0.01800000 0.11525624
#>  [3,] 0.07200000 0.14400000 0.00000000 0.21600000 0.10800000 0.26208396
#>  [4,] 0.10800000 0.03600000 0.21600000 0.00000000 0.03600000 0.07200000
#>  [5,] 0.01800000 0.01800000 0.10800000 0.03600000 0.00000000 0.09000000
#>  [6,] 0.16981166 0.11525624 0.26208396 0.07200000 0.09000000 0.00000000
#>  [7,] 0.16595180 0.13104198 0.24149534 0.10800000 0.09693297 0.01800000
#>  [8,] 0.14400000 0.16099689 0.13104198 0.19134262 0.11384200 0.18356470
#>  [9,] 0.14400000 0.14400000 0.16595180 0.14512064 0.10800000 0.11384200
#> [10,] 0.09693297 0.14058449 0.07200000 0.19386593 0.09000000 0.21600000
#> [11,] 0.10948973 0.18089776 0.00000000 0.25200000 0.14400000 0.28856195
#> [12,] 0.03600000 0.08049845 0.04024922 0.14512064 0.03600000 0.18356470
#> [13,] 0.08049845 0.03600000 0.18089776 0.01800000 0.00000000 0.05091169
#> [14,] 0.06489992 0.05400000 0.14843180 0.05091169 0.01800000 0.07421590
#> [15,] 0.06489992 0.12074767 0.03600000 0.18356470 0.07421590 0.21674870
#> [16,] 0.05400000 0.06489992 0.08049845 0.11384200 0.01800000 0.14512064
#> [17,] 0.07200000 0.10182338 0.06489992 0.15379207 0.05091169 0.18000000
#> [18,] 0.07200000 0.07200000 0.12074767 0.09000000 0.03600000 0.10800000
#>             [,7]       [,8]       [,9]      [,10]     [,11]      [,12]
#>  [1,] 0.16595180 0.14400000 0.14400000 0.09693297 0.1094897 0.03600000
#>  [2,] 0.13104198 0.16099689 0.14400000 0.14058449 0.1808978 0.08049845
#>  [3,] 0.24149534 0.13104198 0.16595180 0.07200000 0.0000000 0.04024922
#>  [4,] 0.10800000 0.19134262 0.14512064 0.19386593 0.2520000 0.14512064
#>  [5,] 0.09693297 0.11384200 0.10800000 0.09000000 0.1440000 0.03600000
#>  [6,] 0.01800000 0.18356470 0.11384200 0.21600000 0.2885620 0.18356470
#>  [7,] 0.00000000 0.14400000 0.07200000 0.18089776 0.2577208 0.16099689
#>  [8,] 0.14400000 0.00000000 0.03600000 0.03600000 0.1018234 0.09000000
#>  [9,] 0.07200000 0.03600000 0.00000000 0.08049845 0.1609969 0.09693297
#> [10,] 0.18089776 0.03600000 0.08049845 0.00000000 0.0180000 0.03600000
#> [11,] 0.25772078 0.10182338 0.16099689 0.01800000 0.0000000 0.07200000
#> [12,] 0.16099689 0.09000000 0.09693297 0.03600000 0.0720000 0.00000000
#> [13,] 0.07200000 0.14058449 0.09693297 0.14843180 0.2160000 0.10800000
#> [14,] 0.06489992 0.10182338 0.07200000 0.10948973 0.1800000 0.07200000
#> [15,] 0.18792552 0.07200000 0.10182338 0.01800000 0.0360000 0.00000000
#> [16,] 0.12074767 0.07200000 0.07200000 0.04024922 0.1080000 0.00000000
#> [17,] 0.14843180 0.05400000 0.06489992 0.00000000 0.0720000 0.01800000
#> [18,] 0.08049845 0.06489992 0.05400000 0.07200000 0.1440000 0.04024922
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.08049845 0.06489992 0.06489992 0.05400000 0.07200000 0.07200000
#>  [2,] 0.03600000 0.05400000 0.12074767 0.06489992 0.10182338 0.07200000
#>  [3,] 0.18089776 0.14843180 0.03600000 0.08049845 0.06489992 0.12074767
#>  [4,] 0.01800000 0.05091169 0.18356470 0.11384200 0.15379207 0.09000000
#>  [5,] 0.00000000 0.01800000 0.07421590 0.01800000 0.05091169 0.03600000
#>  [6,] 0.05091169 0.07421590 0.21674870 0.14512064 0.18000000 0.10800000
#>  [7,] 0.07200000 0.06489992 0.18792552 0.12074767 0.14843180 0.08049845
#>  [8,] 0.14058449 0.10182338 0.07200000 0.07200000 0.05400000 0.06489992
#>  [9,] 0.09693297 0.07200000 0.10182338 0.07200000 0.06489992 0.05400000
#> [10,] 0.14843180 0.10948973 0.01800000 0.04024922 0.00000000 0.07200000
#> [11,] 0.21600000 0.18000000 0.03600000 0.10800000 0.07200000 0.14400000
#> [12,] 0.10800000 0.07200000 0.00000000 0.00000000 0.01800000 0.04024922
#> [13,] 0.00000000 0.00000000 0.14400000 0.07200000 0.10948973 0.04024922
#> [14,] 0.00000000 0.00000000 0.10800000 0.03600000 0.07200000 0.00000000
#> [15,] 0.14400000 0.10800000 0.00000000 0.03600000 0.00000000 0.07200000
#> [16,] 0.07200000 0.03600000 0.03600000 0.00000000 0.00000000 0.00000000
#> [17,] 0.10948973 0.07200000 0.00000000 0.00000000 0.00000000 0.03600000
#> [18,] 0.04024922 0.00000000 0.07200000 0.00000000 0.03600000 0.00000000
#> 
#> attr(,"class")
#> [1] "patch_config"